pyproject.toml
[build-system]

requires = [
    "poetry-core>=2.0.0,<3.0.0"
]

build-backend = "poetry.core.masonry.api"


[project]

name = "strwythura"
description = "Construct a knowledge graph from unstructured data sources, organized by results from entity resolution, implementing an enhanced GraphRAG approach, and also implementing an ontology pipeline plus context engineering for optimizing AI application outcomes within a specific domain."

authors = [
    { name = "Paco Nathan", email = "paco@senzing.com"}
]

readme = "README.md"
license = { text = "MIT" }

version = "1.4.0"

classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Human Machine Interfaces",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Indexing",
]

keywords = [
    "baml",
    "context engineering",
    "entity-linking",
    "entity-resolution",
    "graph algorithms",
    "graphrag",
    "interactive visualization",
    "knowledge-graph",
    "named-entity-recognition",
    "nlp",
    "ontology",
    "ontology-pipeline",
    "relation-extraction",
    "semantic-expansion",
    "semantic-layer",
    "semantic-random-walk",
    "textgraphs",
    "unstructured-data",
    "vector-database",
]

requires-python = ">=3.11,<3.12"

dependencies = [
    "spacy (>=3.8.7,<4.0.0)",
    "gliner (>=0.2.21,<0.3.0)",
    "gliner-spacy (>=0.0.11,<0.0.12)",
    "accelerate (>=1.10.1,<2.0.0)",
    "networkx (>=3.5,<4.0)",
    "rdflib (>=7.1.4,<8.0.0)",
    "lancedb (>=0.24.2,<0.25.0)",
    "polars (>=1.32.3,<2.0.0)",
    "pyvis (>=0.3.2,<0.4.0)",

    "baml-cli (>=0.1.0,<0.2.0)",
    "baml-py (<=0.205.0)",
    "ollama (>=0.5.1,<0.6.0)",
    "gensim (>=4.3.3,<5.0.0)",
    "datasets (>=4.0.0,<5.0.0)",

    "requests (>=2.32.4,<3.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "requests-cache (>=1.2.1,<2.0.0)",
    "w3lib (>=2.3.1,<3.0.0)",
    "lxml (>=6.0.0,<7.0.0)",

    "icecream (>=2.1.5,<3.0.0)",
    "ipywidgets (>=8.1.7,<9.0.0)",
    "jupyterlab (>=4.4.5,<5.0.0)",
    "jupyterlab-execute-time (>=3.2.0,<4.0.0)",
    "pyinstrument (>=5.0.3,<6.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "watermark (>=2.5.0,<3.0.0)",
    "ipython (>=9.4.0,<10.0.0)",
]


[project.urls]

homepage = "https://github.com/DerwenAI/strwythura"
doi = "https://doi.org/10.5281/zenodo.16934079"
slides = "https://derwen.ai/s/2njz#1"
video = "https://senzing.com/gph-graph-rag-llm-knowledge-graphs/"


[tool.poetry.group.dev.dependencies]

pylint = "^3.3.7"
mypy = "^1.17.1"
types-requests = "^2.32.4.20250611"
pandas-stubs = "^2.3.0.250703"
types-networkx = "^3.5.0.20250728"

[[tool.mypy.overrides]]

module = "strwythura.baml_client.*"
ignore_errors = true

[tool.pylint.main]

ignore-paths = [
    "strwythura/baml_client",
]
